fmt: use the real ID_Continue table in FormatValidIdentifier
The local js_lexer stub in fmt.rs returns true for any c > 0x7F, which
the previous cursor.width gate was compensating for. After dropping that
gate, fmt_identifier would pass through non-ID_Continue code points like
U+2014 em dash verbatim. Use crate::string::lexer (the same Unicode table
ensure_valid_identifier uses) so the two copies agree on every input.

The bundled output was already correct for these inputs because the
NumberRenamer re-sanitises via ensure_valid_identifier, but fmt_identifier
itself now honours its contract again. Added a ./foo\u2014bar.cjs case to
the path-derived wrapper test to lock it in.

+++ src/bun_core/fmt.rs
@@ -1234,7 +1234,7 @@ pub struct FormatValidIdentifier<'a> {
 
 impl Display for FormatValidIdentifier<'_> {
     fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        use crate::js_lexer;
+        use crate::string::lexer as js_lexer;
 
         let mut iterator = strings::CodepointIterator::init(self.name);
         let mut cursor = strings::Cursor::default();
@@ -1248,11 +1248,11 @@ impl Display for FormatValidIdentifier<'_> {
         }
 
         // Common case: no gap necessary. No allocation necessary.
-        needs_gap = !js_lexer::is_identifier_start(cursor.c);
+        needs_gap = !js_lexer::is_identifier_start(cursor.c as u32);
         if !needs_gap {
             // Are there any non-alphanumeric chars at all?
             while iterator.next(&mut cursor) {
-                if !js_lexer::is_identifier_continue(cursor.c) {
+                if !js_lexer::is_identifier_continue(cursor.c as u32) {
                     needs_gap = true;
                     start_i = cursor.i as usize;
                     break;
@@ -1274,7 +1274,7 @@ impl Display for FormatValidIdentifier<'_> {
             cursor = strings::Cursor::default();
 
             while iterator.next(&mut cursor) {
-                if js_lexer::is_identifier_continue(cursor.c) {
+                if js_lexer::is_identifier_continue(cursor.c as u32) {
                     if needs_gap {
                         f.write_str("_")?;
                         needs_gap = false;
